Validate Phase Matrix and Calibrations

+++ src/icrar/leap-accelerate/model/cpu/MetaData.cc
@@ -130,7 +130,7 @@ namespace cpu
         casacore::Vector<double> time = msmc->time().getColumn();
         double epoch = time[0];
         int nEpochs = 0;
-        for(int i = 0; i < time.size(); i++)
+        for(size_t i = 0; i < time.size(); i++)
         {
             if(time[i] == time[0]) nEpochs++;
         }
@@ -200,7 +200,7 @@ namespace cpu
         auto size = m_oldUVW.size();
         m_UVW.clear();
         m_UVW.reserve(m_oldUVW.size());
-        for(int n = 0; n < size; n++)
+        for(size_t n = 0; n < size; n++)
         {
             m_UVW.push_back(m_oldUVW[n] * dd);
         }
